pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "dev-code"
dynamic = ["version"]
description = "Project · editor · container — simplified"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.scripts]
devcode = "devcode:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
force-include = {"src/devcode.py" = "devcode.py", "src/templates" = "dev_code_templates"}

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=7.0.1",
    "pyfiglet>=1.0.4",
    "pytest-cov>=4.0",
    "tox>=4.0",
    "pyyaml>=6.0",
]
